// src/api/sqlmodule/model/publicationStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type PublicationStatus = typeof PublicationStatus[keyof typeof PublicationStatus];


export const PublicationStatus = {
  Draft: 'Draft',
  Published: 'Published',
  Archived: 'Archived',
} as const;